+++ java_console/ui/src/main/java/com/rusefi/TriggerWheelInfo.java
@@ -82,6 +82,10 @@ class TriggerWheelInfo {
 
     private boolean isFirstCrankBased() {
         return id == Fields.TT_TT_GM_LS_24 ||
+                id == Fields.TT_TT_HONDA_K_12_1 ||
+                id == Fields.TT_TT_RENIX_44_2_2 ||
+                id == Fields.TT_TT_RENIX_66_2_2_2 ||
+                id == Fields.TT_TT_MIATA_VVT ||
                 id == Fields.TT_TT_GM_7X;
     }