+++ unit_tests/tests/sensor/test_sensor_init.cpp
@@ -51,6 +51,42 @@ TEST(SensorInit, Tps) {
 	EXPECT_NEAR(50.0f, Sensor::get(SensorType::Tps1).value_or(-1), EPS2D);
 }
 
+TEST(SensorInit, TpsValuesTooClose) {
+	WITH_ENGINE_TEST_HELPER(TEST_ENGINE);
+
+	// Should fail, 0.49 volts apart
+	CONFIG(tpsMin) = 200;	// 1.00 volt
+	CONFIG(tpsMax) = 298;	// 1.49 volts
+	EXPECT_FATAL_ERROR(initTps(PASS_CONFIG_PARAMETER_SIGNATURE));
+	Sensor::resetRegistry();
+
+	// Should fail, -0.49 volts apart
+	CONFIG(tpsMin) = 298;	// 1.49 volt
+	CONFIG(tpsMax) = 200;	// 1.00 volts
+	EXPECT_FATAL_ERROR(initTps(PASS_CONFIG_PARAMETER_SIGNATURE));
+	Sensor::resetRegistry();
+
+	// Should succeed, 0.51 volts apart
+	CONFIG(tpsMin) = 200;	// 1.00 volt
+	CONFIG(tpsMax) = 302;	// 1.51 volts
+	EXPECT_NO_FATAL_ERROR(initTps(PASS_CONFIG_PARAMETER_SIGNATURE));
+	Sensor::resetRegistry();
+
+	// Should succeed, -0.51 volts apart
+	CONFIG(tpsMin) = 302;	// 1.51 volt
+	CONFIG(tpsMax) = 200;	// 1.00 volts
+	EXPECT_NO_FATAL_ERROR(initTps(PASS_CONFIG_PARAMETER_SIGNATURE));
+	Sensor::resetRegistry();
+
+	// With no pin, it should be ok that they are the same
+	// Should succeed, -0.51 volts apart
+	CONFIG(tps1_1AdcChannel) = EFI_ADC_NONE;
+	CONFIG(tpsMin) = 200;	// 1.00 volt
+	CONFIG(tpsMax) = 200;	// 1.00 volts
+	EXPECT_NO_FATAL_ERROR(initTps(PASS_CONFIG_PARAMETER_SIGNATURE));
+	Sensor::resetRegistry();
+}
+
 TEST(SensorInit, Pedal) {
 	WITH_ENGINE_TEST_HELPER(TEST_ENGINE);