+++ firmware/CHANGELOG.md
@@ -27,6 +27,10 @@ All notable user-facing or behavior-altering changes will be documented in this 
 
 ## XXX 202X Release
 
+### Added
+ - DAIHATSU 3/4 cylinder trigger https://rusefi.com/images/triggers/trigger_6.png
+ - Ford ST170 VVT cam trigger https://rusefi.com/images/triggers/trigger_42.png
+
 ## December 2020 National Mutt Day
 
 ### Fixed